Moving feedback box out into a context provider
